% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betapriors.R
\name{EB.local}
\alias{EB.local}
\alias{EB}
\title{Empirical Bayes Prior Distribution for Coefficients in BMA Model}
\usage{
EB.local()
}
\value{
returns an object of class "prior", with the family and
hyerparameters.
}
\description{
Creates an object representing the EB prior for BAS GLM.
}
\details{
Creates a structure used for \code{\link{bas.glm}}.
}
\examples{
EB.local()


}
\seealso{
\code{\link{CCH}} and \code{\link{bas.glm}}

Other coef priors: \code{\link{CCH}},
  \code{\link{EB.global}}, \code{\link{Jeffreys}},
  \code{\link{hyper.g.n}}, \code{\link{hyper.g}},
  \code{\link{tCCH}}
}
\author{
Merlise Clyde
}
