\name{EB}
\alias{EB}
\alias{EB.local}
\title{Empirical Bayes Prior Distribution for Coefficients in BMA Model}
\description{Creates an object representing the EB prior  for BAS GLM. 
}
\usage{EB.local()
}

\value{returns an object of class "prior", with the family and
  hyerparameters.
}
\details{Creates a structure used for \code{\link{bas.glm}}. 
    }
\seealso{\code{\link{CCH}} and \code{\link{bas.glm}}}
\examples{EB.local()} 

\author{Merlise Clyde}

