% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_models.R
\name{extract_MPM}
\alias{extract_MPM}
\title{Extract the Median Probability Model}
\usage{
extract_MPM(object)
}
\arguments{
\item{object}{An object of class "bas" or "basglm"}
}
\value{
a new object with of class "bas" or "basglm" with the Median
Probability Model
}
\description{
Extracts the Median Probability Model from a bas object
}
\details{
The Median Probability Model is the model where variables are
included if the marginal posterior probabilty of the coefficient being
zero is greater than 0.5.  As this model may not have been sampled (and even
if it has) it is oftern faster to refit the model using bas, rather than
search the list of models to see where it was included.
}
\examples{
data(Hald, package=BAS)
hald_bic =  bas.lm(Y ~ ., data=Hald, alpha=13, prior="BIC")
extract_MPM(hald_bic)

data(Pima.tr, package="MASS")
Pima_bas = bas.glm(type ~ ., data=Pima.tr, n.models= 2^7, method="BAS",
           betaprior=CCH(a=1, b=nrow(Pima.tr)/2, s=0), family=binomial(),
           modelprior=uniform())
extract_MPM(Pima_bas)
}
\seealso{
Other bas methods: 
\code{\link{BAS}},
\code{\link{bas.lm}()},
\code{\link{coef.bas}()},
\code{\link{confint.coef.bas}()},
\code{\link{confint.pred.bas}()},
\code{\link{diagnostics}()},
\code{\link{fitted.bas}()},
\code{\link{force.heredity.bas}()},
\code{\link{image.bas}()},
\code{\link{plot.confint.bas}()},
\code{\link{predict.basglm}()},
\code{\link{predict.bas}()},
\code{\link{summary.bas}()},
\code{\link{update.bas}()},
\code{\link{variable.names.pred.bas}()}
}
\concept{bas methods}
