% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_priors.R
\name{FND}
\alias{FND}
\title{------------------------------------------------
Cost-Penalized Priors
------------------------------------------------}
\usage{
FND(b = 1)
}
\description{
------------------------------------------------
Cost-Penalized Priors
------------------------------------------------
}
