% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{print.coef.bas}
\alias{print.coef.bas}
\title{Print coefficients generated from coef.bas}
\usage{
\method{print}{coef.bas}(x, digits = max(3, getOption("digits") - 3), ...)
}
\description{
Print coefficients generated from coef.bas
}
