\name{protein}
\docType{data}
\alias{protein}
\title{Protein Activity Data}
\usage{data(protein)}
\description{This data sets includes several predictors of protein
  activity from an experiment run at Glaxo.
}

\format{
  \code{protein} is a dataframe  with 96 observations  and 8 predictor
  variablesof protein activity:
  \tabular{llll}{
  [,1] \tab buf \tab factor \tab  Buffer \cr
  [,2] \tab pH  \tab numeric \tab  \cr
  [,3] \tab NaCl \tab numeric \tab \cr
  [,4] \tab con \tab numeric \tab  protein concentration\cr
  [,5] \tab ra  \tab factor \tab  reducing agent\cr
  [,6] \tab det \tab factor \tab   detergent\cr
  [,7] \tab MgCl2 \tab  numeric\tab \cr
  [,8] \tab temp  \tab numeric\tab  (temerature)\cr
  [,9] \tab prot.act1 \tab numeric\tab \cr
  [,10] \tab prot.act2 \tab numeric\tab \cr
  [,11] \tab prot.act3 \tab numeric\tab \cr
  [,12] \tab prot.act4  \tab numeric\tab protein activity 
}}

\source{Clyde, M. A. and Parmigiani, G. (1998), Protein Construct
  Storage: Bayesian Variable Selection and Prediction with Mixtures,
  Journal of Biopharmaceutical Statistics, 8, 431-443} 


\keyword{datasets}
