% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_priors.R
\name{SBSB}
\alias{SBSB}
\title{------------------------------------------------
New Model Space Priors
------------------------------------------------}
\usage{
SBSB()
}
\description{
------------------------------------------------
New Model Space Priors
------------------------------------------------
}
