% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_priors.R
\name{tCCH}
\alias{tCCH}
\title{Generalized tCCH g-Prior Distribution for Coefficients in BMA Models}
\usage{
tCCH(alpha = 1, beta = 2, s = 0, r = 3/2, v = 1, theta = 1)
}
\arguments{
\item{alpha}{a scalar > 0, recommended alpha=.5 (betaprime) or 1.}

\item{beta}{a scalar > 0.  The value is not updated by the data; beta should
be a function of n for consistency under the null model.}

\item{s}{a scalar, recommended s=0 a priori}

\item{r}{r arbitrary; in the hyper-g-n prior sets r = (alpha + 2)}

\item{v}{0 < v}

\item{theta}{theta > 1}
}
\value{
returns an object of class "prior", with the family and
hyerparameters.
}
\description{
Creates an object representing the tCCH mixture of g-priors on coefficients
for BAS.
}
\details{
Creates a structure used for \code{\link{bas.glm}}.
}
\examples{
n <- 500
tCCH(alpha = 1, beta = 2, s = 0, r = 1.5, v = 1, theta = 1 / n)
}
\seealso{
\code{\link{CCH}}, \code{\link{robust}}, \code{\link{hyper.g}},
\code{\link{hyper.g.n}}\code{\link{bas.glm}}

Other beta priors: 
\code{\link{CCH}()},
\code{\link{EB.local}()},
\code{\link{IC.prior}()},
\code{\link{Jeffreys}()},
\code{\link{TG}()},
\code{\link{beta.prime}()},
\code{\link{g.prior}()},
\code{\link{hyper.g.n}()},
\code{\link{hyper.g}()},
\code{\link{intrinsic}()},
\code{\link{robust}()},
\code{\link{testBF.prior}()}
}
\author{
Merlise Clyde
}
\concept{beta priors}
