% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betapriors.R
\name{testBF.prior}
\alias{testBF.prior}
\title{Test based Bayes Factors for BMA Models}
\usage{
testBF.prior(g)
}
\arguments{
\item{g}{a scalar used in the covariance of Zellner's g-prior, Cov(beta) =
sigma^2 g (X'X)^-}
}
\value{
returns an object of class "prior", with the family and
hyerparameters.
}
\description{
Creates an object representing the prior distribution on coefficients for
BAS that coreesponds to the test-based Bayes Factors.
}
\details{
Creates a prior object structure used for BAS in `bas.glm`.
}
\examples{

testBF.prior(100)
library(MASS)
data(Pima.tr)

# use g = n 
bas.glm(type ~ ., data=Pima.tr, family=binomial(), 
        betaprior=testBF.prior(nrow(Pima.tr)),
        modelprior=uniform(), method="BAS")
}
\seealso{
\code{\link{g.prior}}, \code{\link{bas.glm}}

Other prior functions: \code{\link{IC.prior}},
  \code{\link{TG}}, \code{\link{g.prior}},
  \code{\link{robust}}
}
\author{
Merlise Clyde
}
