% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_priors.R
\name{beta.prime}
\alias{beta.prime}
\title{Beta-Prime Prior Distribution for Coefficients in BMA Model}
\usage{
beta.prime(n = NULL)
}
\arguments{
\item{n}{the sample size; if NULL, the value derived from the data in the
call to `bas.glm` will be used.}
}
\value{
returns an object of class "prior", with the family and
hyerparameters.
}
\description{
Creates an object representing the Beta-Prime prior that is mixture of
g-priors on coefficients for BAS.
}
\details{
Creates a structure used for \code{\link{bas.glm}}.
}
\examples{
beta.prime(n = 100)
}
\seealso{
\code{\link{CCH}}

Other beta priors: 
\code{\link{CCH}()},
\code{\link{EB.local}()},
\code{\link{IC.prior}()},
\code{\link{Jeffreys}()},
\code{\link{TG}()},
\code{\link{g.prior}()},
\code{\link{hyper.g.n}()},
\code{\link{hyper.g}()},
\code{\link{intrinsic}()},
\code{\link{robust}()},
\code{\link{tCCH}()},
\code{\link{testBF.prior}()}
}
\author{
Merlise Clyde
}
\concept{beta priors}
