% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betapriors.R
\name{EB.local}
\alias{EB}
\alias{EB.local}
\title{Empirical Bayes Prior Distribution for Coefficients in BMA Model}
\usage{
EB.local()
}
\value{
returns an object of class "prior", with the family and
hyerparameters.
}
\description{
Creates an object representing the EB prior for BAS GLM.
}
\details{
Creates a structure used for \code{\link{bas.glm}}.
}
\examples{
EB.local()


}
\author{
Merlise Clyde
}
\seealso{
\code{\link{CCH}} and \code{\link{bas.glm}}

Other beta priors: \code{\link{CCH}},
  \code{\link{IC.prior}}, \code{\link{Jeffreys}},
  \code{\link{TG}}, \code{\link{beta.prime}},
  \code{\link{g.prior}}, \code{\link{hyper.g.n}},
  \code{\link{hyper.g}}, \code{\link{intrinsic}},
  \code{\link{robust}}, \code{\link{tCCH}},
  \code{\link{testBF.prior}}
}

