% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betapriors.R
\name{robust}
\alias{robust}
\title{Robust-Prior Distribution for Coefficients in BMA Model}
\usage{
robust(n = NULL)
}
\arguments{
\item{n}{the sample size.}
}
\value{
returns an object of class "prior", with the family and
hyerparameters.
}
\description{
Creates an object representing the robust prior of Bayarri et al (2012) that
is mixture of g-priors on coefficients for BAS.
}
\details{
Creates a structure used for \code{\link{bas.glm}}.
}
\examples{
robust(100)

}
\seealso{
\code{\link{CCH}} and\code{\link{bas.glm}}

Other prior functions: \code{\link{IC.prior}},
  \code{\link{TG}}, \code{\link{g.prior}},
  \code{\link{testBF.prior}}
}
\author{
Merlise Clyde
}
