% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelpriors.R
\name{uniform}
\alias{uniform}
\alias{Uniform}
\title{Uniform Prior Distribution for Models}
\usage{
uniform()
}
\value{
returns an object of class "prior", with the family name Uniform.
}
\description{
Creates an object representing the prior distribution on models for BAS.
}
\details{
The Uniform prior distribution is a commonly used prior in BMA, and is a
special case of the independent Bernoulli prior with probs=.5. The implied
prior distribution on model size is binomial(p, .5).
}
\examples{
uniform()

}
\seealso{
\code{\link{bas.lm}},
\code{\link{beta.binomial}},\code{\link{Bernoulli}},

Other priors modelpriors: \code{\link{Bernoulli.heredity}},
  \code{\link{Bernoulli}}, \code{\link{beta.binomial}},
  \code{\link{tr.beta.binomial}}, \code{\link{tr.poisson}},
  \code{\link{tr.power.prior}}
}
\author{
Merlise Clyde
}
